Moving BIP-340 related keys to key module
